rename catalyst folder to ui-components
